{{cookiecutter.project_slug}}/requirements/dev.txt
-r base.txt
Sphinx>=1.7.1
bumpversion>=0.5.3
coverage>=4.5.1
django-debug-toolbar==1.9.1
factory-boy>=2.10.0
